Implemented FXTools
b1dfdb831f14c7b1362397b3d7bc59ba26556212ed0fc74fbd416b61a1252b46

- FXTools: New utility tool to replace Legacy MFuncs
  - Implemented:
    - `make`
    - `clear`
    - `disk`
    - `qemu`
    - `help`
    - `init`
    - `checksum`
  - Significantly more modular
  - Internal caching system
  - Native chaining system with `+`
  - Uses GRUB bootloader instead of relying on QEMU's built-in emulation.
  - Arch independant, OS independant.
- Architectures store `ARCH_NAME`
- Boots into kernel shell by default
- Removed shelf application
- Fixed GRUB bootloader files

+++ include/farix.h
@@ -71,7 +71,7 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #define SYS_GET_TASK_LIST       SUPER_MIN_SYSCALL + 11
 #define SYS_TASK_KILL           SUPER_MIN_SYSCALL + 12
 
-int32_t farix_syscall (uint32_t sys_id, uint32_t arg1, uint32_t arg2, uint32_t arg3, uint32_t arg4, uint32_t arg5);
+int64_t farix_syscall(uint64_t sys_id, uint64_t arg1, uint64_t arg2, uint64_t arg3, uint64_t arg4, uint64_t arg5);
 
 typedef struct {
     bool isdir;
